moved config file handler to data
